Use Stream and Print classes for input and output
Instead of keeping pin numbers and output handler functions and deciding
on the fly where to send any printed bytes, now a Print* is kept,
greatly simplifying sbp() and friends. Additionally, output can now also
be redirected to an arbitrary Print object by calling setOutputHandler.

Similarly, for the input a Stream* is kept. This allows users of the
library to select any of the available serial port or any other Stream
object to use as the main console. This allows for example to use an
Ethernet client, without needing the bitlash library to understand every
particular ethernet library available (as long as it implements the
Stream interface).

This change breaks compatibility with Arduino versions older than 012,
since those do not have the Print class defined, making it non-trivial
to support those.

This also temporarily breaks compatibility with Arduino versions below
019, which don't contain the Stream class, but this will be fixed in a
subsequent commit.