Don't include Arduino.h in bitlash.cpp
This is already done in bitlash.h, so this is not needed. However, in
bitlash.h we do have to move up the Arduino.h include a bit, because the
code that determines the build type depends on CORE_TEENSY, which is
defined by Arduino.h when compiling for the Teensy.