Clean up support for older Arduino versions
Now, src/bitlash.h requires that either ARDUINO or UNIX_BUILD is
defined. If this is not the case, compilation fails with an error.

Arduino versions below 018 did not define the ARDUINO variable. This
means that for those versions, the user must manually set the Arduino
version to make the build work.

Furthermore, the ARDUINO_BUILD and ARDUINO_VERSION macros are dropped,
since now the ARDUINO variable can just always be used.